move ui data transformation into a separate class
